Obtain lock before obtaining release data
This ensures the release data used is the latest available in
the informer cache once the lock is obtained.
