+++ pkg/release/release.go
@@ -3,8 +3,6 @@ package release
 import (
 	"context"
 	"fmt"
-	"os"
-	"path"
 	"path/filepath"
 	"time"
 
@@ -13,7 +11,6 @@ import (
 
 	corev1client "k8s.io/client-go/kubernetes/typed/core/v1"
 
-	"github.com/fluxcd/helm-operator/internal/lockedfile"
 	"github.com/fluxcd/helm-operator/pkg/apis/helm.fluxcd.io/v1"
 	"github.com/fluxcd/helm-operator/pkg/chartsync"
 	v1client "github.com/fluxcd/helm-operator/pkg/client/clientset/versioned/typed/helm.fluxcd.io/v1"
@@ -70,13 +67,6 @@ func (r *Release) Sync(hr *v1.HelmRelease) (err error) {
 	}
 	logger := releaseLogger(r.logger, client, hr)
 
-	// acquire lock
-	unlock, err := r.lock(fmt.Sprintf("%s-%s", hr.GetNamespace(), hr.GetName()))
-	if err != nil {
-		logger.Log("info", fmt.Sprintf("could not obtain lock: %s", err))
-		return nil
-	}
-	defer unlock()
 	defer func(start time.Time) {
 		ObserveRelease(start, err == nil, hr.GetTargetNamespace(), hr.GetReleaseName())
 	}(time.Now())
@@ -116,12 +106,6 @@ func (r *Release) Sync(hr *v1.HelmRelease) (err error) {
 		return
 	}
 	return r.run(logger, client, action, hr, curRel, chart, values)
-}
-
-func (r *Release) lock(name string) (unlock func(), err error) {
-	lockFile := path.Join(os.TempDir(), name+".lock")
-	mutex := lockedfile.MutexAt(lockFile)
-	return mutex.Lock()
 }
 
 // Uninstalls removes the Helm release for the given HelmRelease,
